Don't force any alignment for the Cent struct
CI tests fail on Ubuntu with LDC 1.30.0 for the compiled 32-bit
binaries. And the culprit is the Cent alignment. The problem is not
reproducible with the older LDC 1.20.1

Looks like all this alignment stuff is rather fragile:
  https://issues.dlang.org/show_bug.cgi?id=22980
  https://github.com/ldc-developers/ldc/issues/1356#issuecomment-1054295164

The 16 byte alignment is not necessary in principle. Except for maybe
some C library bindings, which require compatibility with the '__int128'
type. C compilers don't support '__int128' for 32-bit code though.

For now just use the default struct alignment for Cent. This problem
can be investigated later and the 16 byte alignment may possibly
return at least for the 64-bit binaries.