Implement onelinerizer.rb script
The purpose of this script is to allow converting the whole library
into a compact single line of D code with all comments and extra
whitespaces removed. The unittests can be removed too.

This is useful for submitting code to verious programming contest
websites (such as Codeforces or AtCoder), where normal DUB packages
are not available.