Get rid of pragma inline because it's not reliable after all
In some cases DMD is unable to inline a function and spits out warnings
with the '-wi' option (and this causes a failure at AtCoder):

  Main.d(1): Warning: cannot inline function `zz.Int128.opBinary!"*".opBinary`
  Main.d(1): Warning: cannot inline function `zz.Int128.opOpAssign!"*".opOpAssign`

The documentation at https://dlang.org/spec/pragma.html#inline says that:
"Whether a particular function can be inlined or not is implementation defined."
and "What happens for pragma(inline, true) if the function cannot be inlined.
An error message is typical." This isn't reassuring at all!

So try an alternative solution for the performance loss prevention. DUB supports
"sourceLibrary" as one of the possible "targetType" configurations, as explained
at https://dub.pm/package-format-json (this appears to be somewhat similar to
to LTO or -fwhole-program, because the compiler can optimize across the module
boundaries). Just do it this way for now.